change data type to int

+++ Code/lib/LogicalClock.cpp
@@ -1,22 +1,22 @@
-#include "logicalClock.h"
+#include "LogicalClock.h"
 
 void logicalClock::logicalClockTick()
 {
     this->logicalClock_u64++;
 }
 
-bool logicalClock::logicalClockTickCompare(uint64_t varGlobalLogicalClock)
+bool logicalClock::logicalClockTickCompare(int varGlobalLogicalClock)
 {
     return (this->logicalClock_u64 == varGlobalLogicalClock);
 }
 
-void logicalClock::logicalClockUpdate(uint64_t varGlobalLogicalClock)
+void logicalClock::logicalClockUpdate(int varGlobalLogicalClock)
 {
     this->logicalClock_u64 = (this->logicalClock_u64 > varGlobalLogicalClock) ? (this->logicalClock_u64) : (varGlobalLogicalClock);
     //printf(RED_TEXT "Event: Logical clock differs from Leader! \n" RESET_COLOR);
 }
 
-uint64_t logicalClock::get_logicalClock()
+int logicalClock::get_logicalClock()
 {
     return this->logicalClock_u64;
 }
